pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "submittal-packager"
version = "0.1.0"
description = "INDOT roadway plan submittal packager"
authors = [{name = "AutoGenerated", email = "dev@example.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "typer[all]",
    "pydantic>=1.10",
    "rich>=13",
    "PyPDF2>=3",
    "pdfminer.six>=20221105",
    "jinja2>=3",
    "python-docx>=0.8",
    "pyyaml>=6",
    "tomli>=2",
    "pathspec>=0.11",
    "loguru>=0.7",
]

[project.optional-dependencies]
test = ["pytest", "pytest-cov"]

[project.scripts]
submittal-packager = "submittal_packager.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
